requirements.txt
Django>=4.2.0
python-dotenv>=1.0.0
Pillow>=10.0.0
django-imagekit>=4.1.0  # For thumbnail generation
numpy>=1.24.0  # Required for image processing
imagehash>=4.3.1  # For perceptual image hashing
requests>=2.31.0  # For weather API calls

# Production dependencies
gunicorn>=21.2.0  # Production server
whitenoise>=6.6.0  # Serving static files
django-storages>=1.14.2  # For AWS S3
boto3>=1.34.14  # AWS SDK for Python
psycopg2-binary>=2.9.9  # PostgreSQL adapter
dj-database-url>=2.1.0  # Database URL configuration 